% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My print function}
\usage{
hello(x = "Paulo")
}
\arguments{
\item{x}{Name of a person}
}
\value{
The result of print
}
\description{
My print function
}
\examples{
hello("Paulo")

}
